+++ src/context/index.tsx
@@ -1,0 +1,2 @@
+export { useData } from "../context/dataContext";
+export { useService } from "../context/serviceContext";
